feat: initial diffs system rework
Current `VersionDiff` implementation has a lot of limitations which I have
to face when add a new game. With this set of patches
`VersionDiff` will be transfered into a trait, like it
happened to `Config` struct in the SDK. For each game I'll create
its own `VersionDiff` struct, as well as `DiffUpdate` and `DiffError`.
I believe this will greatly simplify future work, as well as improve overall code quality