feat: predownload completion indication
Change install and preload to use a separate temp subdir for game and
every voicepack. Create a `.predownloadcomplete` file in the temp subdir
when predownload is complete. It gets deleted with the entire contents
of the temp subdir after actually installing.
The launcher can check for that file to know whether predownlaod is
completed. Doesn't account for fails during preload, but those would be
detected and fixed during the actual installation/update